+++ .gitignore
@@ -1,1 +1,4 @@
 .vscode/
+/test/
+*/__pycache__
+__pycache__
