+++ results.tex
@@ -1,0 +1,2 @@
+\newcommand*{\resultTestZweiDrei}{2.0}
+\newcommand*{\resultTest}{1.0}
